Bring back tslint

+++ src/app/manifest/manifest-service.ts
@@ -240,7 +240,7 @@ class ManifestService {
     this.statusText = `${t('Manifest.Load')}...`;
     const currentManifestVersion = localStorage.getItem(this.localStorageKey);
     if (currentManifestVersion === version) {
-      const typedArray = (await get(this.idbKey)) as Uint8Array;
+      const typedArray = await get(this.idbKey);
       if (!typedArray) {
         throw new Error('Empty cached manifest file');
       }
